+++ src/components/Blog/BlogArticleCard/BlogArticleCard.tsx
@@ -2,13 +2,12 @@ import { TrackingAction, TrackingEventParameter } from '@/const/trackingKeys';
 import { JUMPER_LEARN_PATH } from '@/const/urls';
 import { useUserTracking } from '@/hooks/userTracking/useUserTracking';
 import { useMenuStore } from '@/stores/menu/MenuStore';
-import type { StrapiImageData, TagData } from '@/types/strapi';
+import type { BlogArticleData } from '@/types/strapi';
 import { formatDate } from '@/utils/formatDate';
 import { readingTime } from '@/utils/readingTime';
 import type { CSSObject } from '@mui/material';
 import { Skeleton } from '@mui/material';
 import Link from 'next/link';
-import type { RootNode } from 'node_modules/@strapi/blocks-react-renderer/dist/BlocksRenderer';
 import { useTranslation } from 'react-i18next';
 import {
   BlogArticleCardContainer,
@@ -23,34 +22,20 @@ import {
 } from '.';
 
 interface BlogArticleCardProps {
+  article: BlogArticleData;
   baseUrl: string;
-  id: number;
-  image: StrapiImageData;
-  content: RootNode[];
-  publishedAt: string | null | undefined;
-  createdAt: string;
-  title: string;
-  tags: TagData | undefined;
   trackingCategory: string;
-  slug: string;
   styles?: CSSObject;
 }
 
 export const BlogArticleCard = ({
+  article,
   baseUrl,
   trackingCategory,
-  image,
-  tags,
-  content,
-  publishedAt,
-  createdAt,
-  id,
-  title,
   styles,
-  slug,
 }: BlogArticleCardProps) => {
   const { trackEvent } = useUserTracking();
-  const minRead = readingTime(content);
+  const minRead = readingTime(article.attributes.Content);
   const { t } = useTranslation();
   const { closeAllMenus } = useMenuStore((state) => state);
 
@@ -60,15 +45,15 @@ export const BlogArticleCard = ({
       action: TrackingAction.ClickArticleCard,
       label: 'click-blog-article-card',
       data: {
-        [TrackingEventParameter.ArticleTitle]: title,
-        [TrackingEventParameter.ArticleCardId]: id,
+        [TrackingEventParameter.ArticleTitle]: article.attributes.Title,
+        [TrackingEventParameter.ArticleCardId]: article.id,
       },
     });
     closeAllMenus();
   };
   return (
     <Link
-      href={`${JUMPER_LEARN_PATH}/${slug}`}
+      href={`${JUMPER_LEARN_PATH}/${article.attributes.Slug}`}
       style={{ textDecoration: 'none', width: '100%' }}
     >
       <BlogArticleCardContainer
@@ -76,10 +61,13 @@ export const BlogArticleCard = ({
         onClick={handleClick}
         sx={styles}
       >
-        {image.data ? (
+        {article.attributes.Image.data ? (
           <BlogArticleCardImage
-            src={`${baseUrl}${image.data?.attributes?.formats.small.url}`}
-            alt={image.data?.attributes?.alternativeText ?? title}
+            src={`${baseUrl}${article.attributes.Image.data?.attributes?.formats.small.url}`}
+            alt={
+              article.attributes.Image.data?.attributes?.alternativeText ??
+              article.attributes.Title
+            }
             // read the following to udnerstand why width and height are set to 0, https://github.com/vercel/next.js/discussions/18474#discussioncomment-5501724
             width={0}
             height={0}
@@ -100,17 +88,20 @@ export const BlogArticleCard = ({
 
         <BlogArticleCardContent>
           <BlogArticleCardTitle variant="bodyLarge">
-            {title}
+            {article.attributes.Title}
           </BlogArticleCardTitle>
           <BlogArticleCardDetails>
-            {tags?.data.slice(0, 1).map((tag, index) => (
+            {article.attributes.tags?.data.slice(0, 1).map((tag, index) => (
               <BlogArticleCardTag key={index} variant="bodyXSmall" as="h3">
                 {tag.attributes.Title}
               </BlogArticleCardTag>
             ))}
             <BlogArticleCardMetaContainer>
               <BlogArticleMetaDate variant="bodyXSmall" as="span">
-                {formatDate(publishedAt || createdAt)}
+                {formatDate(
+                  article.attributes.publishedAt ||
+                    article.attributes.createdAt,
+                )}
               </BlogArticleMetaDate>
               <BlogArticleMetaReadingTime variant="bodyXSmall" as="span">
                 {t('blog.minRead', { minRead: minRead })}
