+++ src/components/Blog/BlogArticlesCollections/BlogArticlesCollections.style.ts
@@ -1,4 +1,4 @@
-import type { Breakpoint, GridProps } from '@mui/material';
+import type { Breakpoint } from '@mui/material';
 import { Grid } from '@mui/material';
 
 import { styled } from '@mui/material/styles';
@@ -46,36 +46,6 @@ export const BlogArticlesCollectionsContainer = styled(Grid)(({ theme }) => ({
       marginBottom: theme.spacing(12),
     },
   },
-}));
-
-interface ArticlesGridProps extends GridProps {
-  active?: boolean;
-}
-
-export const ArticlesGrid = styled(Grid, {
-  shouldForwardProp: (prop) => prop !== 'active',
-})<ArticlesGridProps>(({ theme, active }) => ({
-  ...(active && {
-    margin: theme.spacing(2, 'auto'),
-    display: 'grid',
-    // marginTop: `calc(${theme.spacing(5)} + ${theme.spacing(8)} + ${theme.spacing(6)} )`, // title height + tabs container + actual offset
-    paddingBottom: 0,
-    gridTemplateColumns: '1fr',
-    justifyItems: 'center',
-    gap: theme.spacing(3),
-    [theme.breakpoints.up('md' as Breakpoint)]: {
-      gridTemplateColumns: '1fr 1fr',
-      gap: theme.spacing(4),
-    },
-    [theme.breakpoints.up('lg' as Breakpoint)]: {
-      gridTemplateColumns: '1fr 1fr 1fr',
-    },
-    [theme.breakpoints.up('xl' as Breakpoint)]: {
-      maxWidth: theme.breakpoints.values.xl,
-      marginLeft: 'auto',
-      marginRight: 'auto',
-    },
-  }),
 }));
 
 export const BlogArticlesCollectionsTitle = styled(CarouselTitle)(
